Create data.txt
this is the example data. It is the litecoin price for certain period taken via coingecko api using similar scripts as discussed in my videos. note the delimiter is a tab not a space! If for some reasons the downloaded file uses spaces, increase the column number in "usecols" by 1!